+++ packages/helpermodules/update_config.py
@@ -32,7 +32,7 @@ from helpermodules.pub import Pub
 from helpermodules.utils.json_file_handler import write_and_check
 from helpermodules.utils.run_command import run_command
 from helpermodules.utils.topic_parser import decode_payload, get_index, get_second_index
-from control.counter_all import counter_all
+from control.counter_all import counter_all, counter_all_data
 from control.bat_all import BatConsiderationMode
 from control.chargepoint.charging_type import ChargingType
 from control.counter import get_counter_default_config
@@ -587,8 +587,8 @@ class UpdateConfig:
         ("openWB/chargepoint/get/power", 0),
         ("openWB/chargepoint/template/0", get_chargepoint_template_default()),
         ("openWB/counter/get/hierarchy", []),
-        ("openWB/counter/config/consider_less_charging", counter_all.Config().consider_less_charging),
-        ("openWB/counter/config/home_consumption_source_id", counter_all.Config().home_consumption_source_id),
+        ("openWB/counter/config/consider_less_charging", counter_all_data.Config().consider_less_charging),
+        ("openWB/counter/config/home_consumption_source_id", counter_all_data.Config().home_consumption_source_id),
         ("openWB/vehicle/0/name", "Standard-Fahrzeug"),
         ("openWB/vehicle/0/color", DEFAULT_COLORS.VEHICLE.value),
         ("openWB/vehicle/0/info", {"manufacturer": None, "model": None}),
